Move description of role-based based behaviour to separate unit test
